+++ class/helios/HeliosSignature.class.php
@@ -1,6 +1,8 @@
 <?php
 class HeliosSignature {
-	
+
+	const HASH_ALGORITHME = "sha256";
+
 	private $xml_starlet_path;
 	
 	public function __construct($xml_starlet_path = false){
@@ -31,7 +33,7 @@ class HeliosSignature {
 		return true;
 	}
 	
-	public function getSha1($xml_content){
+	public function getHash($xml_content){
 		$tmp_file = tempnam("/tmp/", "s2low_xml_");
 		file_put_contents($tmp_file, $xml_content);
 	
@@ -47,8 +49,8 @@ class HeliosSignature {
 		if (! file_exists($c14n_file)){
 			throw new Exception("Impossible de créer le fichier XML canonique $c14n_file");
 		}
-	
-		$result = sha1_file($c14n_file);
+
+		$result = hash_file(self::HASH_ALGORITHME,$c14n_file);
 	
 		return $result;
 	}
@@ -68,13 +70,13 @@ class HeliosSignature {
 				}
 	            foreach($xml->$tag->Bordereau as $bordereau){
 	            	$id[]= strval($bordereau['Id']);
-	            	$hash[] = $this->getSha1($bordereau->asXML());
+	            	$hash[] = $this->getHash($bordereau->asXML());
 	            }
 			}
 			$isBordereau = true;
         } else if( isset( $xml['Id'] ) && !empty($xml['Id'] ) ) {
         		$id[]  = strval($xml['Id']);
-        		$hash[] = $this->getSha1($xml->asXML());
+        		$hash[] = $this->getHash($xml->asXML());
         		$isBordereau = false;
         } else {
 			throw new Exception("Le bordereau du fichier PES ne contient pas d'identifiant valide, ni la balise PESAller : signature impossible");
@@ -89,38 +91,52 @@ class HeliosSignature {
 	}
 
 	public function injectSignature($original_file_path,$signature, $isBordereau){
-		
+
 		$all_signature = explode(",",$signature);
 
 		$domDocument = new DOMDocument();
 		$domDocument->load($original_file_path,LIBXML_PARSEHUGE);
-	
+
+		$signature_raw = [];
+
 		if( $isBordereau ) {
 			$all_bordereau = $domDocument->getElementsByTagName('Bordereau');
 
 			foreach($all_signature as $num_bordereau => $signature) {
 				$signature_1 = base64_decode($signature);
 				$signatureDOM = new DOMDocument();
-				$signatureDOM->loadXML($signature_1);
+				$signatureDOM->loadXML($signature_1,LIBXML_PARSEHUGE);
 				$signature = $signatureDOM->firstChild->firstChild;
-				$cloned = $signature->cloneNode(TRUE);
-				
+
 				$bordereauNode = $all_bordereau->item($num_bordereau);
 
-				$bordereauNode->appendChild($domDocument->importNode($cloned,true));
+				$text_to_replace = "LIBERSIGN_SIGNATURE_NODE_".mt_rand(0,mt_getrandmax());
+				$signature_raw[$text_to_replace] = $signatureDOM->saveXML($signature);
+
+				$textNode = $domDocument->createTextNode($text_to_replace);
+				$bordereauNode->appendChild($textNode);
 			}
-		}
-		else {
+		} else {
 			$signature_1 = base64_decode($signature);
 			$signatureDOM = new DOMDocument();
-			$signatureDOM->loadXML($signature_1);
-            $signature = $signatureDOM->firstChild->firstChild;
-			
-            $rootNode = $domDocument->documentElement;
-            $rootNode->appendChild($domDocument->importNode($signature,true));
+			$signatureDOM->loadXML($signature_1,LIBXML_PARSEHUGE);
+			$signature = $signatureDOM->firstChild->firstChild;
+
+			$rootNode = $domDocument->documentElement;
+
+			$text_to_replace = "LIBERSIGN_SIGNATURE_NODE_".mt_rand(0,mt_getrandmax());
+			$signature_raw[$text_to_replace] = $signatureDOM->saveXML($signature);
+
+			$textNode = $domDocument->createTextNode($text_to_replace);
+			$rootNode->appendChild($textNode);
 		}
 
-		return $domDocument->saveXml();		
+		$result = $domDocument->saveXML();
+
+		foreach($signature_raw as $text_to_replace => $raw_signature){
+			$result = preg_replace("#$text_to_replace#s",$raw_signature,$result);
+		}
+		return $result;
 	}
 	
 }