+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Corrections
 
 - Le mode document papier complémentaire ne retournait pas la bonne valeur. #386
+- Lorsqu'un worker échoue, on ne sort plus du script
 
 
 ## 3.0.13 - 2018-06-28